requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
scikit-image>=0.18.0
tensorboard>=2.7.0
pyiqa>=0.1.7
lpips>=0.1.4
pytorch-msssim>=1.0.0
thop>=0.1.1
imageio>=2.9.0
ftfy>=6.0.0
regex>=2021.0.0
snnTorch>=0.6.0
git+https://github.com/openai/CLIP.git

